requirements.txt
CherryPy==3.2.4
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
itsdangerous==0.23
requests==2.2.1
python-simple-hipchat==0.2
nose==1.3.0
rednose==0.4.1
newrelic==2.14.0.11